问题：上电时Mode_state模式会莫名其妙+1 已解决，改用int型变量，初始值-1

+++ Drive/Device/KEY.h
@@ -2,7 +2,7 @@
 #define _KEY_H
 #include "stm32f10x.h"
 
-extern u8 Mode_state;
+extern int Mode_state;
 void Key_init(void);
 void Key_Scan(void);
